+++ docs/source/changelog.rst
@@ -3,6 +3,21 @@
 ChangeLog
 #########
 
+Version 0.170.0
+===============
+
+- Fix type stub of Python binding
+
+Version 0.169.1
+===============
+
+- Scale connected body transform before serializing to JSON.
+
+Version 0.169.0
+===============
+
+- Add virtual destructor for PyIkFailureAccumulatorBase.
+
 Version 0.168.5
 ===============